css-align-3_dev/xhtml1print/place-content-shorthand-003.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
      <title>CSS Boc Alignment: place-content shorthand - initial value</title>
      <link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez" />
      <link href="http://www.w3.org/TR/css3-align/#propdef-place-content" rel="help" />
      <meta content="Check that place-content's initial value expands to 'align-content' and 'justify-content'." name="assert" />
      <script src="/resources/testharness.js"></script>
      <script src="/resources/testharnessreport.js"></script>
      <style>
          #test {
             align-content: start;
             justify-content: end;
          }
      </style>
  </head>
  <body>
      <div id="log"></div>
      <div id="test"></div>

      <script>
          var div = document.getElementById("test");
          div.setAttribute("style", "place-content: initial")
          var cs = getComputedStyle(div, null);

          test(function() {
              assert_equals(cs.getPropertyValue("place-content"),
                  "normal normal", "place-content computed value");
          }, "place-content_initial_computed_value");

          test(function() {
              assert_equals(cs.getPropertyValue("align-content"),
                  "normal", "place-content specified value for align-content");
          }, "place-content_initial_align-content");

          test(function() {
              assert_equals(cs.getPropertyValue("justify-content"),
                  "normal", "place-content specified value for justify-content");
          }, "place-content_initial_justify-content");
    </script>
  

</body></html>
